backend/pyproject.toml
[project]
name = "telegram-ai-backend"
version = "0.1.0"
description = "Production-ready Telegram AI assistant backend with Vietnamese NLU and Google Sheets workflows."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "redis>=6.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
